+++ src/layout/components/navbar-options.tsx
@@ -1,19 +1,19 @@
-import { useNavigate } from 'react-router-dom';
+import { useState } from "react";
+import { useNavigate } from "react-router-dom";
 
-import useMediaQuery from '@/hooks/useMediaQuery';
-import { Group, MantineTheme, MediaQuery, Text, ThemeIcon, UnstyledButton } from '@mantine/core';
-import {
-    IconArticle, IconCpu, IconLayoutDashboard, IconServer, IconSettings
-} from '@tabler/icons-react';
+import useMediaQuery from "@/hooks/useMediaQuery";
+import { ActionIcon, Group, MantineTheme, Portal, Tooltip, UnstyledButton } from "@mantine/core";
+import { IconCpu, IconLayoutDashboard, IconServer, IconSettings } from "@tabler/icons-react";
 
 interface NavbarOptionProps {
   icon: React.ReactNode;
   label: string;
   onClick?: () => void;
+  active?: boolean;
 }
 
-const NavbarOption: React.FC<NavbarOptionProps> = (props) => {
-  const { icon, label, onClick } = props;
+export const NavbarOption: React.FC<NavbarOptionProps> = (props) => {
+  const { icon, label, onClick, active } = props;
   const { isSmallerThanMd } = useMediaQuery();
   const position = isSmallerThanMd ? "center" : "left";
 
@@ -24,44 +24,49 @@ const NavbarOption: React.FC<NavbarOptionProps> = (props) => {
     borderRadius: theme.radius.sm,
     color: theme.colors.dark[0],
     "&:hover": {
-      backgroundColor: theme.colors.dark[6],
+      backgroundColor: theme.colors.dark[3],
+    },
+    "&[data-active]": {
+      backgroundColor: theme.colors.dark[3],
     },
   });
 
   return (
-    <UnstyledButton sx={sx} onClick={onClick}>
-      <Group position={position}>
-        <ThemeIcon variant="outline">{icon}</ThemeIcon>
-        <MediaQuery smallerThan={"md"} styles={{ display: "none" }}>
-          <Text size={"sm"}>{label}</Text>
-        </MediaQuery>
-      </Group>
-    </UnstyledButton>
+    <Tooltip label={label} position="right" withArrow color={"gray"}>
+      <UnstyledButton sx={sx} onClick={onClick} data-active={active || undefined}>
+        <Group position={position}>
+          <ActionIcon variant="transparent" size={"sm"}>
+            {icon}
+          </ActionIcon>
+        </Group>
+      </UnstyledButton>
+    </Tooltip>
   );
 };
 
 const NavbarOptions = () => {
+  const [active, setActive] = useState(0);
   const navigate = useNavigate();
   const options: NavbarOptionProps[] = [
     {
-      icon: <IconLayoutDashboard size={16} />,
+      icon: <IconLayoutDashboard size={24} />,
       label: "Dashboard",
       onClick: () => navigate("/"),
     },
     {
-      icon: <IconServer size={16} />,
+      icon: <IconServer size={24} />,
       label: "Disks",
       onClick: () => navigate("/disks"),
     },
     {
-      icon: <IconCpu size={16} />,
+      icon: <IconCpu size={24} />,
       label: "Processes",
       onClick: () => {
         navigate("/processes");
       },
     },
     {
-      icon: <IconSettings size={16} />,
+      icon: <IconSettings size={24} />,
       label: "Settings",
       onClick: () => {
         navigate("/settings");
@@ -69,7 +74,17 @@ const NavbarOptions = () => {
     },
   ];
 
-  const navbarOptions = options.map((option) => <NavbarOption {...option} key={option.label} />);
+  const navbarOptions = options.map((option, index) => (
+    <NavbarOption
+      {...option}
+      key={option.label}
+      active={index === active}
+      onClick={() => {
+        setActive(index);
+        option.onClick?.();
+      }}
+    />
+  ));
   return <>{navbarOptions}</>;
 };