+++ src/layout/index.tsx
@@ -1,12 +1,17 @@
 import { Outlet } from "react-router-dom";
 
+import Ellipsis from "@/components/ellipsis";
+import useTheme from "@/hooks/useTheme";
 import Navbar from "@/layout/components/navbar";
-import { AppShell, Card, Space, Styles, useMantineTheme } from "@mantine/core";
+import { AppShell, Styles, useMantineTheme } from "@mantine/core";
 
 import AppFooter from "./components/footer";
+import Header from "./components/header";
 
 const Layout: React.FC = () => {
   const theme = useMantineTheme();
+  const { isMidnight } = useTheme();
+
   const styles: Styles<"body" | "main" | "root", never> | undefined = {
     main: {
       background: theme.colors.dark[8],
@@ -14,11 +19,9 @@ const Layout: React.FC = () => {
   };
 
   return (
-    <AppShell styles={styles} navbar={<Navbar />} footer={<AppFooter />} padding={"xl"}>
-      <Space h={"sm"} />
-      <Card style={{ backgroundColor: theme.colors.dark[7] }} radius={"md"}>
-        <Outlet />
-      </Card>
+    <AppShell styles={styles} navbar={<Navbar />} footer={<AppFooter />} header={<Header />} padding={"xl"}>
+      {isMidnight && <Ellipsis right="0px" width="75%" blur="150px" top="100" />}
+      <Outlet />
     </AppShell>
   );
 };